+++ components/selfupdater/googleplay/.gitignore
