+++ components/selfupdater/googleplay/src/main/AndroidManifest.xml
