--- components/selfupdater/source/github/src/main/AndroidManifest.xml
@@ -1,4 +1,0 @@
-<manifest package="com.flipperdevices.selfupdater.source.github"
-        xmlns:android="http://schemas.android.com/apk/res/android">
-    <uses-permission android:name="android.permission.REQUEST_INSTALL_PACKAGES"/>
-</manifest>
