--- components/selfupdater/source/github/src/main/kotlin/com/flipperdevices/selfupdater/source/github/model/GithubUpdate.kt
@@ -1,7 +1,0 @@
-package com.flipperdevices.selfupdater.source.github.model
-
-data class GithubUpdate(
-    val version: String,
-    val downloadUrl: String,
-    val name: String,
-)
