--- components/selfupdater/source/unknown/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="com.flipperdevices.selfupdater.source.noop" />
