+++ instances/app/build.gradle.kts
@@ -169,13 +169,18 @@ dependencies {
     implementation(projects.components.selfupdater.api)
     when (SOURCE_INSTALL) {
         SourceInstall.GOOGLE_PLAY -> {
-            implementation(projects.components.selfupdater.source.googleplay)
+            implementation(projects.components.selfupdater.googleplay)
         }
         SourceInstall.GITHUB -> {
-            implementation(projects.components.selfupdater.source.github)
+            implementation(projects.components.selfupdater.thirdparty.api)
+            implementation(projects.components.selfupdater.thirdparty.github)
+        }
+        SourceInstall.FDROID -> {
+            implementation(projects.components.selfupdater.thirdparty.api)
+            implementation(projects.components.selfupdater.thirdparty.fdroid)
         }
         else -> {
-            implementation(projects.components.selfupdater.source.unknown)
+            implementation(projects.components.selfupdater.unknown)
         }
     }
