+++ instances/app/src/main/java/com/flipperdevices/app/FlipperApplication.kt
@@ -25,7 +25,7 @@ class FlipperApplication : Application() {
                 application = this,
                 ApplicationParams(
                     startApplicationClass = SingleActivity::class,
-                    version = BuildConfig.VERSION_NAME
+                    version = "1.3.9"
                 )
             )
