+++ settings.gradle.kts
@@ -189,8 +189,9 @@ include(
     ":components:faphub:installation:impl",
 
     ":components:selfupdater:api",
-    ":components:selfupdater:source:github",
-    ":components:selfupdater:source:googleplay",
-    ":components:selfupdater:source:unknown",
-
+    ":components:selfupdater:googleplay",
+    ":components:selfupdater:unknown",
+    ":components:selfupdater:thirdparty:api",
+    ":components:selfupdater:thirdparty:github",
+    ":components:selfupdater:thirdparty:fdroid",
 )
